backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
prophet==1.1.5
scikit-learn==1.3.2
joblib==1.3.2
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
stripe==7.8.0
